Cut up and de-included mailbox.hpp
