+++ src/clustering/administration/issues/memory.hpp
@@ -6,6 +6,7 @@
 
 #include "clustering/administration/issues/issue.hpp"
 #include "containers/incremental_lenses.hpp"
+#include "rpc/connectivity/server_id.hpp"
 #include "rpc/mailbox/typed.hpp"
 
 class local_issue_server_t;
