+++ src/clustering/administration/logs/log_transfer.cc
@@ -5,6 +5,7 @@
 
 #include "concurrency/promise.hpp"
 #include "containers/archive/boost_types.hpp"
+#include "rpc/mailbox/disconnect_watcher.hpp"
 
 RDB_IMPL_SERIALIZABLE_1_SINCE_v1_13(log_server_business_card_t, address);
