+++ src/clustering/administration/logs/logs_backend.hpp
@@ -14,6 +14,9 @@ class cluster_directory_metadata_t;
 class log_message_t;
 class log_server_business_card_t;
 class server_config_client_t;
+class server_id_t;
+
+template <class key_t, class value_t> class watchable_map_t;
 
 /* This backend assumes that the entries in the log file have timestamps that are unique
 and monotonically increasing. These assumptions can be broken if the system clock runs
@@ -25,6 +28,10 @@ these assumptions are broken, the system shouldn't crash, but the contents of
     runs `r.db("rethinkdb").table("logs").get(primary_key)`
   * Changefeeds on `rethinkdb.logs` might skip some changes.
 */
+
+namespace logs_backend {
+class cfeed_machinery_t;
+}
 
 class logs_artificial_table_backend_t :
     public cfeed_artificial_table_backend_t
@@ -63,62 +70,7 @@ public:
             admin_err_t *error_out);
 
 private:
-    class cfeed_machinery_t : public cfeed_artificial_table_backend_t::machinery_t {
-    public:
-        cfeed_machinery_t(
-            namespace_id_t const &namespace_id,
-            lifetime_t<name_resolver_t const &> name_resolver,
-            auth::user_context_t const &user_context,
-            logs_artificial_table_backend_t *_parent);
-
-        /* `on_change()` checks for newly-connected peers. If it finds one, it puts an
-        entry in `peers_handled` and spawns an instance of `run()`. */
-        void on_change(const peer_id_t &peer, const cluster_directory_metadata_t *dir);
-
-        /* One instance of `run` will be running for each server we're in contact with
-        that hasn't been permanently removed. It first fetches the latest entry of each
-        server's log, then repeatedly checks for newer log entries at a regular interval.
-        If it sees that the server is disconnected, then it removes itself from
-        `peers_handled` and stops. */
-        void run(
-            const peer_id_t &peer,
-            const server_id_t &server_id,
-            const log_server_business_card_t &bcard,
-            bool is_a_starter,
-            auto_drainer_t::lock_t keepalive);
-
-        /* Helper function for `run()`. Checks if the server is no longer present in the
-        directory; if so, removes the entry from `peers_handled` and returns `false`. */
-        bool check_disconnected(const peer_id_t &peer);
-
-        /* Called by `cfeed_artificial_table_backend_t` to fetch the initial values for a
-        new changefeed */
-        bool get_initial_values(
-            const new_mutex_acq_t *proof,
-            std::vector<ql::datum_t> *initial_values_out,
-            signal_t *interruptor);
-
-        logs_artificial_table_backend_t *parent;
-        std::set<peer_id_t> peers_handled;
-        std::map<peer_id_t, timespec> last_timestamps;
-
-        /* `all_starters_done` is pulsed when we've fetched logs from every peer that was
-        connected to the `cfeed_machinery_t` when it was first created. When the
-        `cfeed_machinery_t` is first created, `starting` is true, and any instance of
-        `run()` that are spawned in the first group have `is_a_starter` set to `true`.
-        `num_starters_left` is initially the number of such coroutines. As soon as the
-        initial batch are spawned, `starting` is set to `false`, so any further instances
-        that are spawned for newly-connected servers will have `is_a_starter` set to
-        `false`. As each instance with `is_a_starter` set to `true` finishes fetching the
-        initial timestamp, it decrements `num_starters_left`. The last one pulses
-        `all_starters_done`.*/
-        bool starting;
-        int num_starters_left;
-        cond_t all_starters_done;
-
-        auto_drainer_t drainer;
-        watchable_map_t<peer_id_t, cluster_directory_metadata_t>::all_subs_t dir_subs;
-    };
+    friend class logs_backend::cfeed_machinery_t;
 
     bool read_all_rows_raw(
         const std::function<void(
