+++ src/clustering/administration/servers/server_metadata.hpp
@@ -7,6 +7,7 @@
 #include <string>
 
 #include "containers/name_string.hpp"
+#include "rpc/connectivity/server_id.hpp"
 #include "rpc/mailbox/typed.hpp"
 #include "rpc/semilattice/joins/deletable.hpp"
 #include "rpc/semilattice/joins/macros.hpp"
