+++ src/clustering/generic/minidir.tcc
@@ -2,6 +2,7 @@
 #include "clustering/generic/minidir.hpp"
 
 #include "concurrency/wait_any.hpp"
+#include "rpc/mailbox/mailbox.hpp"
 
 template<class key_t, class value_t>
 minidir_read_manager_t<key_t, value_t>::minidir_read_manager_t(
