+++ src/clustering/generic/registrar.hpp
@@ -9,10 +9,14 @@
 
 #include "clustering/generic/registration_metadata.hpp"
 #include "concurrency/coro_pool.hpp"
+#include "concurrency/mutex.hpp"
 #include "concurrency/promise.hpp"
 #include "concurrency/queue/unlimited_fifo.hpp"
 #include "concurrency/wait_any.hpp"
+#include "containers/map_sentries.hpp"
+#include "rpc/mailbox/disconnect_watcher.hpp"
 #include "rpc/mailbox/typed.hpp"
+#include "utils.hpp"
 
 template<class business_card_t, class user_data_type, class registrant_type>
 class registrar_t {
