+++ src/clustering/immediate_consistency/backfillee.cc
@@ -2,6 +2,7 @@
 #include "clustering/immediate_consistency/backfillee.hpp"
 
 #include "arch/timing.hpp"
+#include "assignment_sentry.hpp"
 #include "clustering/immediate_consistency/history.hpp"
 #include "concurrency/cross_thread_signal.hpp"
 #include "concurrency/wait_any.hpp"
