+++ src/unittest/fifo_enforcer.cc
@@ -8,6 +8,7 @@
 #include "concurrency/auto_drainer.hpp"
 #include "concurrency/fifo_enforcer_queue.hpp"
 #include "concurrency/wait_any.hpp"
+#include "random.hpp"
 #include "unittest/unittest_utils.hpp"
 #include "unittest/gtest.hpp"
