+++ src/unittest/mock_store.cc
@@ -2,6 +2,7 @@
 
 #include "arch/timing.hpp"
 #include "btree/backfill.hpp"
+#include "random.hpp"
 #include "rdb_protocol/store.hpp"
 
 namespace unittest {
